move logic into api
